requirements.txt
requests
paramiko
networkx
matplotlib
reportlab
python-dotenv
openai
pymetasploit3
